api_UCT/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
python-dotenv==1.0.1
strawberry-graphql==0.246.1
